finish

+++ src/entrypoint.c
@@ -6,7 +6,7 @@
 /*   By: vide-sou <vide-sou@student.42.rio>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/15 15:21:48 by vide-sou          #+#    #+#             */
-/*   Updated: 2024/11/22 09:53:51 by vide-sou         ###   ########.fr       */
+/*   Updated: 2024/11/22 11:38:00 by vide-sou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 static char	*join_all_stdin(int argc, char **argv)
 {
-	size_t	index;
+	int		index;
 	char	*tmp;
 	char	*result;
